refactor: use features for schema and client
The client module now contains only code for hitting the osv API
endpoint.

The rust bindings to the osv schema are now the only thing enabled
by default and now live in the schema module.

fixes: #7.
