+++ examples/parse.rs
@@ -2,8 +2,10 @@ use std::env;
 use std::fs;
 use std::process;
 
+use osv::client;
+
 #[async_std::main]
-async fn main() -> Result<(), osv::ApiError> {
+async fn main() -> Result<(), client::ApiError> {
     let args: Vec<String> = env::args().skip(1).collect();
     if args.len() <= 0 {
         println!("filename expected");
@@ -12,7 +14,7 @@ async fn main() -> Result<(), osv::ApiError> {
     for arg in &args {
         let path: &str = arg.as_str();
         let file = fs::File::open(path).unwrap();
-        let _vuln: osv::Vulnerability =
+        let _vuln: osv::schema::Vulnerability =
             serde_json::from_reader(file).unwrap_or_else(|_| panic!("fail: {}", path));
         println!("pass: {}", path);
     }
