+++ examples/vulnerability.rs
@@ -1,6 +1,8 @@
+use osv::client;
+
 #[async_std::main]
-async fn main() -> Result<(), osv::ApiError> {
-    let vuln = osv::vulnerability("GHSA-jfh8-c2jp-5v3q").await?;
+async fn main() -> Result<(), client::ApiError> {
+    let vuln = client::vulnerability("GHSA-jfh8-c2jp-5v3q").await?;
     println!("{:#?}", vuln);
     Ok(())
 }
